% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arctools.R
\docType{package}
\name{arctools-package}
\alias{arctools}
\alias{arctools-package}
\title{arctools: processing and physical activity summaries of minute-level activity data}
\description{
arctools provides `activity_stats()` function to process minute level
actigraphy-measured activity counts
data and extract commonly used physical activity volume and fragmentation metrics.
}
\details{
To learn more about arctools, start with the vignettes:
`browseVignettes(package = "arctools")`
}
\author{
\strong{Maintainer}: Marta Karas \email{marta.karass@gmail.com} (\href{https://orcid.org/0000-0001-5889-3970}{ORCID})

Authors:
\itemize{
  \item Jennifer Schrack (\href{https://orcid.org/0000-0001-9244-9267}{ORCID})
  \item Jacek Urbanek (\href{https://orcid.org/0000-0002-1890-8899}{ORCID})
}

}
