% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{extdata_fnames}
\alias{extdata_fnames}
\title{Names of exemplary accelerometry data file.}
\format{
A character vector.
}
\usage{
extdata_fnames
}
\description{
Names of exemplary accelerometry data files attached to the package.
These data can be accessed via
\code{system.file("extdata", "<data_file_name>.csv", package = "arctools")}.
}
\keyword{datasets}
