% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_actigraph_SN.R
\name{get_actigraph_SN}
\alias{get_actigraph_SN}
\title{Get Actigraph device serial number}
\usage{
get_actigraph_SN(fpath_full)
}
\arguments{
\item{fpath_full}{A string scalar. An absolute path to ActiLife accelerometry
data file.}
}
\value{
String scalar. Actigraph device serial number.
}
\description{
Read Actigraph device serial number from ActiLife accelerometry data
file.
}
\examples{
fpath_full_i <- system.file("extdata", extdata_fnames[1], package = "arctools")
get_actigraph_SN(fpath_full_i)

}
